Refactor the representation and parsing of variable declarations.
Introduces separate representations in the parsing and typechecking/codegen
passes - the parser just treats a singular definition as an inline statement,
while a translation step checks the scoping rules and translates into explicit
scopes starting with all variables defined in this scope. Old rules in CodeGen
to rename variables is removed and will be replaced by an AST rewriting step in
the ScopeCheck pass.